refactor(news): move bloc.dart to data
