+++ lib/news/widgets/article_preview.dart
@@ -5,7 +5,7 @@ import 'package:hpi_flutter/news/widgets/article_page.dart';
 import 'package:provider/provider.dart';
 import 'package:timeago/timeago.dart' as timeago;
 
-import '../bloc.dart';
+import '../data/bloc.dart';
 
 @immutable
 class ArticlePreview extends StatelessWidget {
